feat(bootstrap.py): Added guard for action file overwrite

+++ bootstrap.py
@@ -82,7 +82,8 @@ def _chroot_to_mnt():
     scripts = glob.glob('*.sh') + glob.glob('*.py')
     for script in scripts:
         shutil.copy(script, MOUNTPOINT)
-    shutil.copy(common.EXECUTED_ACTIONS_FILENAME, MOUNTPOINT)
+    if not os.path.isfile(f'{MOUNTPOINT}/{common.EXECUTED_ACTIONS_FILENAME}'):
+        shutil.copy(common.EXECUTED_ACTIONS_FILENAME, MOUNTPOINT)
     os.chroot(MOUNTPOINT)
     os.chdir('/')
